Done with trade section

+++ WatchdogWebInterface/Models/ExportRecord.cs
@@ -5,10 +5,10 @@ public class ExportRecord
     public int Id { get; set; }
     public string CommodityId { get; set; } = string.Empty;
     public string CommodityDescription { get; set; } = string.Empty;
-    public int TotalValueForMonth { get; set; }
-    public int VesselValueForMonth { get; set; }
-    public int AirValueForMonth { get; set; }
-    public int CardCount { get; set; }
+    public long TotalValueForMonth { get; set; }
+    public long VesselValueForMonth { get; set; }
+    public long AirValueForMonth { get; set; }
+    public long CardCount { get; set; }
     public int Month { get; set; }
     public int Year { get; set; }
